dev/requirements.txt
astor==0.8.1
astroid==3.1.0
black==24.4.0
click==8.1.7
darker==2.1.1
darkgraylib==1.2.0
dill==0.3.8
django-debug-toolbar==4.3.0
djhtml==3.0.6
et-xmlfile==1.1.0
flynt==1.0.1
graylint==1.1.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
toml==0.10.2
tomlkit==0.12.4
